+++ src/main/java/org/postgres/debugger/breakpoint/PlpgLineBreakPointProperties.java
@@ -1,6 +1,0 @@
-package org.postgres.debugger.breakpoint;
-
-import com.intellij.database.debugger.SqlLineBreakpointProperties;
-
-public class PlpgLineBreakPointProperties extends SqlLineBreakpointProperties {
-}