+++ src/main/java/org/postgres/debugger/command/PlpgTask.java
@@ -1,9 +1,0 @@
-package org.postgres.debugger.command;
-
-public enum PlpgTask {
-    DUMMY,
-    STEP_INTO,
-    STEP_OVER,
-    ABORT,
-    RESUME
-}