+++ src/main/java/org/postgres/debugger/command/result/PlpgVoidResult.java
@@ -1,4 +1,0 @@
-package org.postgres.debugger.command.result;
-
-public class PlpgVoidResult extends PlpgDebuggerResult {
-}